balena-python/requirements.txt
hologram-python==0.10.1
pyroute2==0.5.*
pyserial~=3.5
python-pppd==1.0.4
python-sdk-auth==0.4.0
pyudev~=0.22.0
pyusb~=1.2.1
psutil~=5.8.0
requests>=2.25.1
